+++ lib/src/commands/generate.dart
@@ -9,10 +9,6 @@ import '../utils/exception.dart';
 
 class GenerateCommand extends FaradayCommand {
   GenerateCommand() : super() {
-    argParser.addOption('ios-common', help: 'ios FaradayCommon.swift');
-    argParser.addOption('ios-route', help: 'ios FaradayRoute.swift');
-    argParser.addOption('android-common', help: 'android FaradayCommon.kt');
-    argParser.addOption('android-route', help: 'android FaradayRoute.kt');
     argParser.addOption('file', abbr: 'f', help: '解析指定文件');
   }
 
@@ -73,18 +69,17 @@ class GenerateCommand extends FaradayCommand {
     final configPath = path.join(root, '.faraday.json');
     final config = JSON.parse(File(configPath).readAsStringSync());
 
-    final ios_common = stringArg('ios-common') ?? config['ios-common'].string;
-    final ios_route = stringArg('ios-route') ?? config['ios-route'].string;
-    final android_common =
-        stringArg('android-common') ?? config['android-common'].string;
-    final android_route =
-        stringArg('android-route') ?? config['android-route'].string;
+    final ios_common = config['ios-common'].string;
+    final ios_route = config['ios-route'].string;
+    final android_common = config['android-common'].string;
+    final android_route = config['android-route'].string;
 
     return <String, String>{
       if (ios_common != null) 'ios-common': ios_common,
       if (ios_route != null) 'ios-route': ios_route,
       if (android_common != null) 'android-common': android_common,
       if (android_route != null) 'android-route': android_route,
+      'dart-route': path.join(root, 'lib/src/routes.dart')
     };
   }
 }